Interactive configuration.
Began to write a partial configuration in an interactive way through the
JoyceApplication. There, also the mapping download is included, however
there is not yet a service actually doing the download and also it is
currently not clear how to create the aggregates within Joyce (without
the need to use the Semedico Neo4j database).